Update visualizing-with-bokeh.md
Various suggested edits as well as syntax and grammar tweaks to aid readability and translation. @archaeocharlie: please check you are happy with these changes and confirm. Note this is part 1 of my edits (up to 'Categorical Data and Bar Charts: Munitions Dropped by Country'). I'll continue to work on this today.