Updates part I, fixes and extends code
@maltevogl and I worked on slimming down Part I and rewrote it to offer a better flow from the original historical research interest to the introduction of simulation methods / abm as a possible answer to the questions posed. A big part of the more complex topics surrounding epsitemology and methodology were pushed to Part III for now and will be heavily reworked / worked into there.

We aligned and partially revised the structure of the draft compared to the original submission in order to facilitate the changes made in Part I.

We also fixed the code and extended it to include an interactable visualization in jupyter if the reader interacts with it there. This can also be converted into a static image later for demsontration in the lesson itself.